requirements.txt
numpy==1.26.4
fastapi
pydantic
uvicorn
transformers
spaces
pillow
requests
timm
einops
optimum
accelerate
opencv-python
